Howls [clean-arch]: Application Business Layer
- implement use cases layer, the business rules layer of clean architecture

Use Cases: Use Cases is the second domain layer. It defines application-specific business rules. They encapsulate and implement all of the approved use cases for the application. The use cases control the flow to and from entities and can call on entities to use their enterprise-scale rules to complete certain user tasks.

Changes in this layer will not affect entities or more external layers. However, this layer will have to be changed if external layers are changed.